Scale temp down to match reality

+++ src/main/java/dynamodb/TemperatureItem.java
@@ -16,7 +16,7 @@ public class TemperatureItem {
     private String id;
 
     @DynamoDBAttribute(attributeName = DynamoAttributes.TEMPERATURE_ATTRIBUTE_NAME)
-    private int temperature;
+    private double temperature;
 
     @DynamoDBAttribute(attributeName = DynamoAttributes.SENSOR_ADDRESS_ATTRIBUTE_NAME)
     private String sensorAddress;
@@ -24,7 +24,7 @@ public class TemperatureItem {
     @DynamoDBAttribute(attributeName = DynamoAttributes.RECEIVE_TIME_ATTRIBUTE_NAME)
     private long receiveTime;
 
-    public TemperatureItem(int temperature, String sensorAddress) {
+    public TemperatureItem(double temperature, String sensorAddress) {
         this.id = UUID.randomUUID().toString();
         this.temperature = temperature;
         this.sensorAddress = sensorAddress;
@@ -42,11 +42,11 @@ public class TemperatureItem {
         this.id = id;
     }
 
-    public int getTemperature() {
+    public double getTemperature() {
         return temperature;
     }
 
-    public void setTemperature(int temperature) {
+    public void setTemperature(double temperature) {
         this.temperature = temperature;
     }
